Enemy Formations txt should now work fine in builds of the game, and health pickup spawn positions were fixed

+++ game/Roboquatic/Roboquatic/HealthPickup.cs
@@ -18,7 +18,7 @@ namespace Roboquatic
 
         public HealthPickup(Rectangle position, Game1 game)
         {
-            this.position = new Rectangle((int)(position.X), (int)(position.Y), (int)(position.Width * GlobalScalars.x), (int)(position.Height * GlobalScalars.y));
+            this.position = new Rectangle(position.X, position.Y, (int)(position.Width * GlobalScalars.x), (int)(position.Height * GlobalScalars.y));
             pickupSprite = game.Content.Load<Texture2D>("UpgradeHealth");
         }
